+++ conman/routes/exceptions.py
@@ -1,0 +1,2 @@
+class InvalidURL(Exception):
+    """Raised when a url does not appear to be valid."""